% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actimetric.R
\docType{package}
\name{actimetric}
\alias{actimetric}
\title{actimetric: A package for applying machine learning models to accelerometer data.}
\description{
The actimetric package provides functions to apply machine learning models to
accelerometer data for extracting physical activity and sleep-related features
}
\section{actimetric functions}{

The actimetric functions ...
}

