% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actimetricPlot.R
\name{actimetricPlot}
\alias{actimetricPlot}
\title{Generates plots from the time series classified}
\usage{
actimetricPlot(ts, daysummary, dsnames, classes, epoch)
}
\arguments{
\item{ts}{Time series with the activity classified to be stored.}

\item{daysummary}{Matrix with the day-level summary for this recording.}

\item{dsnames}{Names of the columns in the daysummary matrix.}

\item{classes}{Expected classes classified in the time series.}

\item{epoch}{Numeric with epoch length in seconds.}
}
\value{
Does not return any object, it just draws the plots.
}
\description{
Generates plots from the time series classified
}
\author{
Jairo H. Migueles <jairo@jhmigueles.com>
}
