% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_per_person.R
\name{aggregate_per_person}
\alias{aggregate_per_person}
\title{Aggregate Per Recording}
\usage{
aggregate_per_person(
  daysummary,
  n_valid_hours = 0,
  n_valid_hours_awake = 0,
  n_valid_hours_nighttime = 0
)
}
\arguments{
\item{daysummary}{Data frame generated in \link{aggregate_per_date}.}

\item{n_valid_hours}{Numeric (default = 0) with absolute number of valid hours required to consider a day valid}

\item{n_valid_hours_awake}{Numeric (default = 0) with absolute number of awake valid hours required to consider a day valid}

\item{n_valid_hours_nighttime}{Numeric (default = 0) with absolute number of nighttime valid hours required to consider a day valid}
}
\value{
Data frame with aggregates of time spent in classes per recording.
}
\description{
Generates person/recording-level reports
}
