% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattcalibrate.R, R/mattcalibrate_VH.R
\name{calibrate}
\alias{calibrate}
\title{Title}
\usage{
calibrate(file, temp = FALSE, sphere = 0.3, Fs, chunksize = 1, mintime = 84)

calibrate(file, temp = FALSE, sphere = 0.3, Fs, chunksize = 1, mintime = 84)
}
\arguments{
\item{Fs}{}

\item{mintime}{}
}
\description{
Title

Title
}
