% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateRaw.R
\name{calibrateRaw}
\alias{calibrateRaw}
\title{Calibrates Raw Accelerometer Data}
\usage{
calibrateRaw(raw, sf, verbose = TRUE)
}
\arguments{
\item{raw}{Matrix containing accelerometer signal and, if available, temperature data.}

\item{sf}{Sampling frequency in Hertz.}

\item{verbose}{Logical (default = TRUE) indicating whether progress messages should
be printed in the console}
}
\value{
List with calibration coefficients and calibrated raw data
}
\description{
Calibrates the raw accelerometer signal based on the algorithm published
in
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
