% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_classifier.R
\name{check_classifier}
\alias{check_classifier}
\title{Checks the Spelling and Selects the Correct Classifier}
\usage{
check_classifier(classifier)
}
\arguments{
\item{classifier}{Character with user-specified classifier.}
}
\value{
Character with the definitive classifier
}
\description{
Tries to match the classifier specified by the user with one of the
supported options in the package. If not successful returns NULL.
}
\author{
Jairo H. Migueles <jairo@jhmigueles.com>
check_classifier("Adult wrist Trost")
}
