% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkIndexing.R
\name{chunkIndexing}
\alias{chunkIndexing}
\title{Derives Indices to Read Data in Chunks}
\usage{
chunkIndexing(prevChunk = 0, sf = NULL, rawEnd = NULL)
}
\arguments{
\item{prevChunk}{Integer (default = 0) with the chunk index read in the
previous iteration.}

\item{sf}{Numeric (default = NULL) indicating the sampling frequency in Hertz.}

\item{rawEnd}{Numeric (default = NULL) indicating \code{nrow(raw)}.}
}
\value{
Indices to be read in current iteration
}
\description{
Function used to derive indices to read raw data in 24-hour chunks.
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
