% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFClassifier.R
\name{classifier_Preschool}
\alias{classifier_Preschool}
\title{Title}
\usage{
classifier_Preschool(
  raw,
  Fs,
  ID,
  mypath,
  sleep,
  win,
  rfmodel,
  Classifier,
  start.time
)
}
\arguments{
\item{start.time}{}
}
\description{
Title
}
