% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThighClassifier.R
\name{classifier_thigh}
\alias{classifier_thigh}
\title{Title}
\usage{
classifier_thigh(
  raw,
  Fs,
  ID,
  mypath,
  win,
  sleep,
  rfmodel = rfmodel,
  Classifier = Classifier,
  start.time
)
}
\arguments{
\item{start.time}{}
}
\description{
Title
}
