% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrostAdultRFClassifier.R
\name{classifier_TrostAdult}
\alias{classifier_TrostAdult}
\title{Title}
\usage{
classifier_TrostAdult(
  raw,
  Fs,
  ID,
  mypath,
  win,
  sleep,
  rfmodel = rfmodel,
  Classifier = Classifier,
  start.time = start.time
)
}
\arguments{
\item{start.time}{}
}
\description{
Title
}
