% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwa.R
\name{cwa}
\alias{cwa}
\title{Read CWA files}
\usage{
cwa(file, start = 0, end = 0, progressBar = FALSE, desiredtz = c())
}
\arguments{
\item{file}{Character with full path to accelerometer raw data file}

\item{start}{Numeric with starting page for reading file.}

\item{end}{Numeric with ending page for reading file.}

\item{progressBar}{Logical indicating whether to show a progress bar.}

\item{desiredtz}{Desired time zone.}
}
\value{
List with header and datar frame with acceleration data.
}
\description{
Read CWA files
}
