% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveTimestamps.R
\name{deriveTimestamps}
\alias{deriveTimestamps}
\title{Derives timestamps by epoch}
\usage{
deriveTimestamps(from, length, epoch, tz = "")
}
\arguments{
\item{from}{Numeric with starting time for timestamps in UTC format.}

\item{length}{Numeric with the length of the desired timestamp.}

\item{epoch}{Numeric with epoch length in seconds.}

\item{tz}{A character string specifying the time zone to be used for the conversion.
Examples include `"UTC"`, `"America/New_York"`, or `"Europe/Berlin"`.
If not specified, the system's default time zone is used. Time zone handling affects
how character or numeric inputs are interpreted and displayed.
A full list of time zone identifiers can be found on
[Wikipedia](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones).}
}
\value{
Timestamp in "%Y-%m-%d %H:%M:%OS" format
}
\description{
Derives timestamps by epoch
}
\author{
Jairo H. Migueles <jairo@jhmigueles.com>

Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
