% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_sleep_periods.R,
%   R/detect_sleep_periods_new.R
\name{detect_sleep_periods}
\alias{detect_sleep_periods}
\title{Title}
\usage{
detect_sleep_periods(data, window, sleepnr)

detect_sleep_periods(data, window, sleepnr)
}
\arguments{
\item{window}{}

\item{sleepnr}{}
}
\description{
Title

Title
}
