% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectNonWear.R
\name{detectNonWear}
\alias{detectNonWear}
\title{Detects Non Wear Periods}
\usage{
detectNonWear(data, sf, epoch, sdThreshold = 13/1000)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{sf}{Numeric indicating the sampling frequency in Hertz.}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sdThreshold}{Threshold to define nonwear time in g units (default 13 mgs).}
}
\value{
Numeric vector of length = nrow(data) indicating wear (0) and nonwear (1).
}
\description{
Function to detect non-wear periods based on the SD of the vector magnitude
of the raw accelerations. In brief, it looks for 30-min windows in which the
SD of the vector magnitude is lower than a given threshold (default = 0.013 G
units). Additionally, if short non-wear periods (max 30 min) are surrounded with
at least 30% of their duration by non-wear periods, then these short wear periods
are converted to non-wear (additional non-wear).
}
\references{
GGIR R Package. doi: 10.1123/jmpb.2018-0063

van Hees VT, Renström F, Wright A, Gradmark A, Catt M, Chen KY,
Löf M, Bluck L, Pomeroy J, Wareham NJ, Ekelund U, Brage S, Franks PW.
Estimation of daily energy expenditure in pregnant and non-pregnant women
using a wrist-worn tri-axial accelerometer. PLoS One. 2011;6(7):e22922.
doi: 10.1371/journal.pone.0022922
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
