% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSleep.R
\name{detectSleep}
\alias{detectSleep}
\title{Detects Sleep Periods from Wrist Attachment Site}
\usage{
detectSleep(ts, anglez, epoch, sleep_id, nonwear_id)
}
\arguments{
\item{ts}{Data frame with ts object from \link{extractFeatures}}

\item{anglez}{Data frame with timestamp and angle z as calculated in \link{classify}}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sleep_id}{Identification number for sleep periods.}

\item{nonwear_id}{Identification number for nonwear periods.}
}
\value{
Data frame with time series with sleep period indicator included.
}
\description{
Detects Sleep Periods from Wrist Attachment Site
}
