% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSleepThigh.R
\name{detectSleepThigh}
\alias{detectSleepThigh}
\title{Detects Sleep Periods from Thigh Attachment Site}
\usage{
detectSleepThigh(ts, epoch, sleep_id, nonwear_id)
}
\arguments{
\item{ts}{Data frame with ts object from \link{extractFeatures}}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sleep_id}{Identification number for sleep periods.}

\item{nonwear_id}{Identification number for nonwear periods.}
}
\description{
Detects Sleep Periods from Thigh Attachment Site
}
