% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSleepWrist.R
\name{detectSleepWrist}
\alias{detectSleepWrist}
\title{Detects Sleep Periods from Wrist Attachment Site}
\usage{
detectSleepWrist(data, ts, epoch, sf, start_time, sleep_id, nonwear_id)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{ts}{Data frame with ts object from \link{extractFeatures}}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sf}{Numeric with the sampling frequency in Hz.}

\item{start_time}{Start time.}

\item{sleep_id}{Identification number for sleep periods.}

\item{nonwear_id}{Identification number for nonwear periods.}
}
\description{
Detects Sleep Periods from Wrist Attachment Site
}
