% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doReport.R
\name{doReport}
\alias{doReport}
\title{Generates Reports}
\usage{
doReport(files2summarize)
}
\arguments{
\item{files2summarize}{List of files to be summarized in reports.}
}
\value{
No returns, it only saves csv files in results folder.
}
\description{
Generates day-level and person-level reports#'
}
