% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{ExtractFeatures}
\alias{ExtractFeatures}
\title{Extracts Features from raw data}
\usage{
ExtractFeatures(data, classifier = NULL, sf = NULL, epoch = NULL, ID = NA)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{classifier}{Character specifying the classifier to be used
(available options are:
Preschool Wrist Random Forest Free Living,
Preschool Hip Random Forest Free Living,
Preschool Hip Random Forest Free Living Lag-Lead,
Preschool Wrist Random Forest Free Living Lag-Lead,
School age Wrist Random Forest, School age Hip Random Forest,
Adult Wrist RF Trost,
Adult women Wrist RF Ellis,
Adult women Hip RF Ellis)}

\item{sf}{Numeric with the sampling frequency in Hz.}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{ID}{Character with ID corresponding to subject}
}
\value{
Data frame with features calculated as needed for each classifier.
}
\description{
Extracts Features from raw data
}
\details{
This is a central function to calculate features from the acceleration signal.
If any of the implemented classifiers is to be used, then the function
automatically calculates the features needed for such classifier.
}
