% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature extraction.R
\name{feature.extraction}
\alias{feature.extraction}
\title{Title}
\usage{
feature.extraction(
  x = trial2,
  domain = 3,
  vector.mag = 3,
  window = 10,
  overlap = 0,
  freq = 30,
  lower = 0.5,
  upper = 5
)
}
\arguments{
\item{upper}{}
}
\description{
Title
}
