% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features}
\alias{features}
\title{Title}
\usage{
features(
  input,
  output,
  win = 15,
  axis = "yaxis",
  Sed = 200,
  Mod = 420,
  Vig = 842,
  Calibrate = F,
  visual = T,
  Guide = T,
  sleep = T,
  dothis = dothis,
  ndays = 12,
  folder_name = "",
  file_name = ""
)
}
\arguments{
\item{file_name}{}
}
\description{
Title
}
