% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresThigh.R
\name{featuresThigh}
\alias{featuresThigh}
\title{Calculates Features for Thigh classifier}
\usage{
featuresThigh(x, y, z, epoch, sf)
}
\arguments{
\item{x}{Numeric vector with the raw acceleration for X axis in G units.}

\item{y}{Numeric vector with the raw acceleration for Y axis in G units.}

\item{z}{Numeric vector with the raw acceleration for Z axis in G units.}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sf}{Numeric with the sampling frequency in Hz.}
}
\value{
Matrix with features per epoch as proposed in Ellis 2016.
}
\description{
Calculates Features for Thigh classifier
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
