% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresTrost2017.R
\name{featuresTrost2017}
\alias{featuresTrost2017}
\title{Calculates Features Proposed By Trost 2017}
\usage{
featuresTrost2017(data, epoch, sf)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sf}{Numeric with the sampling frequency in Hz.}
}
\value{
Matrix with features per epoch as proposed in Trost 2018.
}
\description{
Calculates Features Proposed By Trost 2017
}
\references{
Pavey TG, Gilson ND, Gomersall SR, Clark B, Trost SG. Field
evaluation of a random forest activity classifier for wrist-worn accelerometer
data. J Sci Med Sport. 2017 Jan;20(1):75-80. doi: 10.1016/j.jsams.2016.06.003

Adult Wrist RF Trost: Pavey TG, Gilson ND, Gomersall SR, Clark B,
Trost SG. Field evaluation of a random forest activity classifier for wrist-worn
accelerometer data. J Sci Med Sport. 2017 Jan;20(1):75-80.
doi: 10.1016/j.jsams.2016.06.003
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>

Stewart G. Trost <s.trost@uq.edu.au>
}
