% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresTrost2018.R
\name{featuresTrost2018}
\alias{featuresTrost2018}
\title{Calculates Features Proposed By Trost 2018}
\usage{
featuresTrost2018(
  data,
  vm,
  epoch,
  overlap,
  sf,
  lowerBound = 0.25,
  upperBound = 5
)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{vm}{Numeric with vector magnitude of the raw accelerations.}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{overlap}{Whether to overlap windows in the features calculation.}

\item{sf}{Numeric with the sampling frequency in Hz.}

\item{lowerBound}{Lower bound for the frequency filter in Hz.}

\item{upperBound}{Upper bound for the frequency filter in Hz.}
}
\value{
Matrix with features per epoch as proposed in Trost 2018.
}
\description{
Calculates Features Proposed By Trost 2018
}
\references{
Trost SG, Cliff DP, Ahmadi MN, Tuc NV, Hagenbuchner M.
Sensor-enabled Activity Class Recognition in Preschoolers: Hip versus Wrist
Data. Med Sci Sports Exerc. 2018 Mar;50(3):634-641. doi: 10.1249/MSS.0000000000001460

Preschool models: Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364

School age Wrist Random Forest: Chowdhury AK, Tjondronegoro D,
Chandran V, Trost SG. Ensemble Methods for Classification of Physical Activities
from Wrist Accelerometry. Med Sci Sports Exerc. 2017 Sep;49(9):1965-1973.
doi: 10.1249/MSS.0000000000001291

School age Hip Random Forest: Ahmadi MN, Pfeiffer KA, Trost SG.
Physical Activity Classification in Youth Using Raw Accelerometer Data from
the Hip. Meas Phys Edu Exerc Sci. 2020; 24(2):129-136. doi: 10.1080/1091367X.2020.1716768
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
