% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genplotSTLaurencewrist.R, R/genplotenmo.R
\name{GenPlot}
\alias{GenPlot}
\title{Title}
\usage{
GenPlot(
  final,
  window = win,
  output,
  folder_name = folder_name,
  file_name = file_name
)

GenPlot(
  final,
  window = win,
  output,
  folder_name = folder_name,
  file_name = file_name
)
}
\arguments{
\item{file_name}{}
}
\description{
Title

Title
}
