% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genplotellis.R
\name{GenPlot_Ellis}
\alias{GenPlot_Ellis}
\title{Title}
\usage{
GenPlot_Ellis(
  final,
  window = win,
  output,
  folder_name = folder_name,
  file_name = file_name
)
}
\arguments{
\item{file_name}{}
}
\description{
Title
}
