% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlocksizeReadaccfile.R
\name{getBlocksizeReadaccfile}
\alias{getBlocksizeReadaccfile}
\title{Get Blocksize for Reading File in Chunks}
\usage{
getBlocksizeReadaccfile(file, sf, monc, dformat)
}
\arguments{
\item{file}{Character with file path to raw acceleration file}

\item{sf}{Numeric with sampling frequency in Hertz}

\item{monc}{Code for monitor as extracted from \code{GGIR::g.inspectfile}}

\item{dformat}{Code for file format as extracted from \code{GGIR::g.inspectfile}}
}
\value{
Block size for reading chunk of data
}
\description{
Get Blocksize for Reading File in Chunks
}
