% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBout.R
\name{getBout}
\alias{getBout}
\title{Calculate bouts of behavior}
\usage{
getBout(x, boutduration = 10, boutcriter = 0.8, boutmaxgap = 1, epoch = 5)
}
\arguments{
\item{x}{Binary numeric vector (default = NULL) with 1's indicating occurrence of behavior.}

\item{boutduration}{Integer (default = 10) indicating the length of the bout in minutes.}

\item{boutcriter}{Numeric (default = 0.8). If -1, then no gaps within the bout are allowed.
If a number lower than 1, then it is considered as a proportion of the boutduration (e.g., 0.8 means 80%).
If 1 or larger, then the gap is interpreted in absolute number of minutes (e.g., 2 means 2 minutes).}

\item{boutmaxgap}{Integer (default = 1) with maximum consecutive gap length allowed in bout calculation.}

\item{epoch}{Integer (default = 5) indicating the epoch length in seconds.}
}
\description{
Calculates bouts of the different behaviors provided by the classifier
}
