% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInfoClassifier.R
\name{GetInfoClassifier}
\alias{GetInfoClassifier}
\title{Get Information To Apply Classifiers}
\usage{
GetInfoClassifier(classifier)
}
\arguments{
\item{classifier}{Character specifying the classifier to be used
(available options are:
Preschool Wrist Random Forest Free Living,
Preschool Hip Random Forest Free Living,
Preschool Hip Random Forest Free Living Lag-Lead,
Preschool Wrist Random Forest Free Living Lag-Lead,
School age Wrist Random Forest, School age Hip Random Forest,
Adult Wrist RF Trost,
Adult women Wrist RF Ellis,
Adult women Hip RF Ellis)}
}
\value{
List containing epoch, rfmodel, and hmmmodel
}
\description{
Gets the required information to run a classifier (i.e., epoch and models)
}
\references{
Preschool models: Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364

School age Wrist Random Forest: Chowdhury AK, Tjondronegoro D,
Chandran V, Trost SG. Ensemble Methods for Classification of Physical Activities
from Wrist Accelerometry. Med Sci Sports Exerc. 2017 Sep;49(9):1965-1973.
doi: 10.1249/MSS.0000000000001291

School age Hip Random Forest: Ahmadi MN, Pfeiffer KA, Trost SG.
Physical Activity Classification in Youth Using Raw Accelerometer Data from
the Hip. Meas Phys Edu Exerc Sci. 2020; 24(2):129-136. doi: 10.1080/1091367X.2020.1716768

Adult Wrist RF Trost: Pavey TG, Gilson ND, Gomersall SR, Clark B,
Trost SG. Field evaluation of a random forest activity classifier for wrist-worn
accelerometer data. J Sci Med Sport. 2017 Jan;20(1):75-80.
doi: 10.1016/j.jsams.2016.06.003

Adult women Wrist and Hip RF Ellis: Ellis K, Kerr J, Godbole S,
Staudenmayer J, Lanckriet G. Hip and Wrist Accelerometer Algorithms for
Free-Living Behavior Classification. Med Sci Sports Exerc. 2016 May;48(5):933-40.
doi: 10.1249/MSS.0000000000000840
}
\author{
Jairo H. Migueles <jairo@jhmigueles.com>
}
