% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep detection.R
\name{inbed}
\alias{inbed}
\title{Title}
\usage{
inbed(
  angle,
  k = 60,
  perc = 0.1,
  inbedthreshold = 15,
  bedblocksize = 30,
  outofbedsize = 60,
  ws3 = 5
)
}
\arguments{
\item{ws3}{}
}
\description{
Title
}
