% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonwear_1min.R
\name{nonwear_vm}
\alias{nonwear_vm}
\title{Title}
\usage{
nonwear_vm(data, Fs, window)
}
\arguments{
\item{window}{}
}
\description{
Title
}
\examples{
if wear period is less than 30 min and less than 30\% of bordering non-wear period,
convert to nonwear
}
