% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.activpal.R
\name{read.activpal}
\alias{read.activpal}
\title{Reads activPAL raw acceleration files}
\usage{
read.activpal(file)
}
\arguments{
\item{file}{Character with path to raw acceleration file}
}
\value{
Data frame with acceleration data
}
\description{
Reads activPAL raw acceleration files
}
