% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_counts.R
\name{read.counts}
\alias{read.counts}
\title{Reads counts files}
\usage{
read.counts(file)
}
\arguments{
\item{file}{Character with path to raw acceleration file}
}
\value{
List containing counts data and epoch size
}
\description{
Reads counts files
}
