% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read excel files}
\usage{
read_excel_allsheets(file, tibble = FALSE)
}
\arguments{
\item{file}{Character with path to file}

\item{tibble}{Read tidyverse tibbles}
}
\value{
List with a data frame per excel sheet.
}
\description{
Read excel files
}
