% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAccFile.R
\name{ReadAccFile}
\alias{ReadAccFile}
\title{Read Accelerometer Raw Data File}
\usage{
ReadAccFile(file, verbose = TRUE)
}
\arguments{
\item{file}{Character with full path to accelerometer raw data file}

\item{verbose}{Logical (default = TRUE) indicating whether progress messages should
be printed in the console.}
}
\value{
List containing header info, accelerometer raw data, start time and ID
for the file.
}
\description{
Function aimed to read accelerometer raw data. At the moment,
the function supports ActiGraph gt3x and csv data, GENEActiv bin
and csv data
}
