% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readActigraph.R
\name{readActigraph}
\alias{readActigraph}
\title{Reads ActiGraph Files}
\usage{
readActigraph(file, verbose = TRUE)
}
\arguments{
\item{file}{Character with full path to actigraph file to read}

\item{verbose}{Logical (default = TRUE) indicating whether progress messages should
be printed in the console.}
}
\value{
List containing header info, accelerometer raw data, start time, and
sampling frequency for the measurement
}
\description{
Function to read ActiGraph Files (support gt3x+, wgt3xBT, and Link devices)
}
