% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndCalibrate.R
\name{ReadAndCalibrate}
\alias{ReadAndCalibrate}
\title{Read Accelerometer Raw Data File and Extract Calibration Coefficients}
\usage{
ReadAndCalibrate(
  file,
  sf,
  blocksize,
  blocknumber,
  inspectfileobject,
  PreviousEndPage,
  PreviousLastValue,
  PreviousLastTime,
  isLastBlock,
  do.calibration,
  iteration,
  epoch,
  S,
  tz = "",
  verbose
)
}
\arguments{
\item{file}{Path to the file to be read.}

\item{sf}{Number with the sampling frequency in the recording.}

\item{blocksize}{Size of chunk to be read in pages as extracted from \code{GGIR::get_nw_clip_block_params}.}

\item{blocknumber}{Number of chunk of data to be read.}

\item{inspectfileobject}{Inspect file object extracted from \code{GGIR::g.inspectfile}.}

\item{PreviousEndPage}{Last page number read in the previous iteration.}

\item{PreviousLastValue}{Last acceleration value(s) read in the previous iteration.}

\item{PreviousLastTime}{Last timestamp read in the previous iteration.}

\item{isLastBlock}{Logical indicating if this is the last chunk of data to be read in the file.}

\item{do.calibration}{Logical indicating whether to calibrate the data or not.}

\item{iteration}{Number indicating the iteration in the \code{runActimetric} function.}

\item{epoch}{Number with the desired epoch length for the aggregation in seconds.}

\item{S}{Leftover data from the previous iteration, to be appended to the current chunk of data being read.}

\item{tz}{A character string specifying the time zone to be used for the conversion.
Examples include `"UTC"`, `"America/New_York"`, or `"Europe/Berlin"`.
If not specified, the system's default time zone is used. Time zone handling affects
how character or numeric inputs are interpreted and displayed.
A full list of time zone identifiers can be found on
[Wikipedia](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones).}

\item{verbose}{Logical indicating whether to print progress messages in the console.}
}
\value{
List containing header info, accelerometer raw data, start time and ID
for the file.
}
\description{
Function aimed to read accelerometer raw data. At the moment,
the function supports ActiGraph gt3x and csv data, GENEActiv bin
and csv data
}
\references{
GGIR R Package. doi: 10.1123/jmpb.2018-0063
}
\author{
Jairo H. Migueles <jairo@jhmigueles.com>

Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
