% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readax3.R
\name{readax3}
\alias{readax3}
\title{Reads Axivity AX3 files}
\usage{
readax3(file)
}
\arguments{
\item{file}{Character with path to raw acceleration file}
}
\value{
List with sampling frequency, raw acceleration data, and start time.
}
\description{
Reads Axivity AX3 files
}
