% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCsv.R
\name{readCsv}
\alias{readCsv}
\title{Reads CSV Files With Accelerometer Data}
\usage{
readCsv(file)
}
\arguments{
\item{file}{Character with full path to accelerometer raw data file}
}
\value{
List containing header info, accelerometer raw data, sampling frequency,
monitor brand, and start time for the measurement
}
\description{
Reads CSV Files With Accelerometer Data
}
