% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGA.R
\name{readGeneactiv}
\alias{readGeneactiv}
\title{Reads GENEActiv files}
\usage{
readGeneactiv(file)
}
\arguments{
\item{file}{Character with path to raw acceleration file}
}
\value{
Data frame with acceleration data and start time for the measurement.
}
\description{
Reads GENEActiv files
}
