% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runActimetric.R
\name{runActimetric}
\alias{runActimetric}
\title{Classifies Accelerometer Data into Physical Activity Types}
\usage{
runActimetric(
  input_directory = NULL,
  output_directory = NULL,
  studyname = "actimetric",
  do.calibration = TRUE,
  do.sleep = TRUE,
  do.nonwear = TRUE,
  do.enmo = TRUE,
  do.actilifecounts = FALSE,
  do.actilifecountsLFE = FALSE,
  classifier = NULL,
  boutdur = c(10),
  boutcriter = 0.8,
  boutmaxgap = 1,
  n_valid_hours = 0,
  n_valid_hours_awake = 0,
  n_valid_hours_nighttime = 0,
  visualreport = FALSE,
  overwrite = FALSE,
  tz = "",
  verbose = TRUE
)
}
\arguments{
\item{input_directory}{Character with full path to directory containing raw
accelerometer data. At the moment the package accepts ActiGraph gt3x,
axivity cwa, and GENEActiv bin files. The \code{input_directory} should only
contain accelerometer files, they can be organized in subdirectories within
the input directory if needed.}

\item{output_directory}{Character with full path to directory where the reports
and visualizations are to be stored.}

\item{studyname}{Character (default = "actimetric") defining the name of the study.
To be used to give a name to the output directory generated.}

\item{do.calibration}{Logical (default = TRUE) indicating whether the accelerometer
raw signal should be calibrated relative to local gravitational acceleration.}

\item{do.sleep}{Logical (default = TRUE) indicating whether sleep parameters
should be calculated.}

\item{do.nonwear}{Logical (default = TRUE) indicating whether non-wear time
should be derived.}

\item{do.enmo}{Logical (default = TRUE) indicating whether ENMO should be
calculated.}

\item{do.actilifecounts}{Logical (default = FALSE) indicating whether activity
counts should be calculated.}

\item{do.actilifecountsLFE}{Logical (default = FALSE) indicating whether activity
counts using the low-frequency extension filter should be calculated.}

\item{classifier}{Character (default = NULL) indicating the classifier to be used
(available options are:
Preschool Wrist Random Forest Free Living,
Preschool Hip Random Forest Free Living,
Preschool Hip Random Forest Free Living Lag-Lead,
Preschool Wrist Random Forest Free Living Lag-Lead,
School age Wrist Random Forest, School age Hip Random Forest,
Adult Wrist RF Trost,
Adult women Wrist RF Ellis,
Adult women Hip RF Ellis)}

\item{boutdur}{Numeric vector (default = c(10)) indicating the bout durations over which calculate bouts of behaviors}

\item{boutcriter}{Numeric (default = 0.8) indicating the proportion of the bout duration that should be classified in a given behavior to consider a bout}

\item{boutmaxgap}{Integer (default = 1) with maximum consecutive gap length allowed in bout calculation.}

\item{n_valid_hours}{Numeric (default = 0) with minimum number of absolute valid hours in the day to consider it a valid day for the person-level aggregates.}

\item{n_valid_hours_awake}{Numeric (default = 0) with minimum number of absolute valid awake hours in the day to consider it a valid day for the person-level aggregates.}

\item{n_valid_hours_nighttime}{Numeric (default = 0) with minimum number of absolute valid nighttime hours in the day to consider it a valid day for the person-level aggregates.}

\item{visualreport}{Logical (default = FALSE) indicating whether visualizations should be run and stored in the results folder.}

\item{overwrite}{Logical (default = FALSE) indicating whether the existing milestone
data should be overwritten.}

\item{tz}{A character string specifying the time zone to be used for the conversion.
Examples include `"UTC"`, `"America/New_York"`, or `"Europe/Berlin"`.
If not specified, the system's default time zone is used. Time zone handling affects
how character or numeric inputs are interpreted and displayed.
A full list of time zone identifiers can be found on
[Wikipedia](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones).}

\item{verbose}{Logical (default = TRUE) indicating whether progress messages should
be printed in the console.}
}
\value{
Function does not return anything, it only generates the reports and
visualizations in the \code{output_directory}.
}
\description{
This is the central function of the actimetric package.
}
\references{
GGIR R Package. doi: 10.1123/jmpb.2018-0063

Preschool models: Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364

School age Wrist Random Forest: Chowdhury AK, Tjondronegoro D,
Chandran V, Trost SG. Ensemble Methods for Classification of Physical Activities
from Wrist Accelerometry. Med Sci Sports Exerc. 2017 Sep;49(9):1965-1973.
doi: 10.1249/MSS.0000000000001291

School age Hip Random Forest: Ahmadi MN, Pfeiffer KA, Trost SG.
Physical Activity Classification in Youth Using Raw Accelerometer Data from
the Hip. Meas Phys Edu Exerc Sci. 2020; 24(2):129-136. doi: 10.1080/1091367X.2020.1716768

Adult Wrist RF Trost: Pavey TG, Gilson ND, Gomersall SR, Clark B,
Trost SG. Field evaluation of a random forest activity classifier for wrist-worn
accelerometer data. J Sci Med Sport. 2017 Jan;20(1):75-80.
doi: 10.1016/j.jsams.2016.06.003

Adult women Wrist and Hip RF Ellis: Ellis K, Kerr J, Godbole S,
Staudenmayer J, Lanckriet G. Hip and Wrist Accelerometer Algorithms for
Free-Living Behavior Classification. Med Sci Sports Exerc. 2016 May;48(5):933-40.
doi: 10.1249/MSS.0000000000000840

van Hees VT, Gorzelniak L, Dean León EC, Eder M, Pias M, Taherian S,
Ekelund U, Renström F, Franks PW, Horsch A, Brage S. Separating movement and
gravity components in an acceleration signal and implications for the assessment
of human daily physical activity. PLoS One. 2013 Apr 23;8(4):e61691.
doi: 10.1371/journal.pone.0061691

Neishabouri A, Nguyen J, Samuelsson J, Guthrie T, Biggs M, Wyatt J,
Cross D, Karas M, Migueles JH, Khan S, Guo CC. Quantification of acceleration
as activity counts in ActiGraph wearable. Sci Rep. 2022 Jul 13;12(1):11958.
doi: 10.1038/s41598-022-16003-x
}
\author{
Jairo H. Migueles <jairo@jhmigueles.com>

Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
