% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.entropy.r
\name{shannon.entropy}
\alias{shannon.entropy}
\title{Function to compute entropy}
\usage{
shannon.entropy(p)
}
\arguments{
\item{p}{Numeric vector of values from th escaled periodogram generated by the fft function.}
}
\value{
Numeric with shanon entropy
}
\description{
Function to compute entropy
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
