% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick.R
\name{tick_to_posix}
\alias{tick_to_posix}
\title{Converts tick to posix}
\usage{
tick_to_posix(x, tz = "UTC", ...)
}
\arguments{
\item{x}{Tick}

\item{tz}{Time zone}

\item{...}{Additional arguments for as.POSIXct.}
}
\value{
POSIXct object
}
\description{
Converts tick to posix
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
