% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPreschool.R
\name{classifier_Preschool}
\alias{classifier_Preschool}
\title{Title}
\usage{
classifier_Preschool(
  raw,
  sf,
  ID,
  mypath,
  sleep,
  epoch,
  rfmodel,
  classifier,
  start_time
)
}
\arguments{
\item{start_time}{}
}
\description{
Title
}
