% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySleep.R
\name{classifySleep}
\alias{classifySleep}
\title{Classifies nighttime and sleep in the time series}
\usage{
classifySleep(
  anglez,
  starttime,
  classifier,
  infoClassifier,
  ts,
  do.sleep,
  do.nonwear
)
}
\arguments{
\item{anglez}{Angle for the z axis relative to the horizontal plane.}

\item{starttime}{Start time as exported from \link{ReadAndCalibrate}}

\item{classifier}{Character (default = NULL) indicating the classifier to be used
(available options are:
Preschool Wrist Random Forest Free Living,
Preschool Hip Random Forest Free Living,
Preschool Hip Random Forest Free Living Lag-Lead,
Preschool Wrist Random Forest Free Living Lag-Lead,
School age Wrist Random Forest, School age Hip Random Forest,
Adult Wrist RF Trost,
Adult women Wrist RF Ellis,
Adult women Hip RF Ellis)}

\item{infoClassifier}{Information relative to the selected classifier as exported from \link{GetInfoClassifier}}

\item{ts}{Data frame with time series.}

\item{do.sleep}{Logical (default = TRUE) indicating if sleep should be calculated}

\item{do.nonwear}{Logical (default = TRUE) indicating if nonwear should be calculated}
}
\value{
Numeric vector with classified activities including nighttime and sleep.
}
\description{
Function to classify nighttime and sleep in the time series.
}
