% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveTimestamps.R
\name{deriveTimestamps}
\alias{deriveTimestamps}
\title{Derives timestamps by epoch}
\usage{
deriveTimestamps(from, length, epoch)
}
\arguments{
\item{from}{Numeric with starting time for timestamps in UTC format.}

\item{length}{Numeric with the length of the desired timestamp.}

\item{epoch}{Numeric with epoch length in seconds.}
}
\value{
Timestamp in "%Y-%m-%d %H:%M:%OS" format
}
\description{
Derives timestamps by epoch
}
