% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_sleep_periods.R
\name{detect_sleep_periods}
\alias{detect_sleep_periods}
\title{Detects Sleep Periods}
\usage{
detect_sleep_periods(data, epoch, sleep_id)
}
\arguments{
\item{data}{Data frame with time series of data as calculated in \code{detectSleepWrist}}

\item{epoch}{Numeric with epoch size}

\item{sleep_id}{Identificator of sleep periods.}
}
\value{
Indicator of sleep periods.
}
\description{
Function not intended for use by the users. Detects sleep periods for
wrist-worn accelerometers.
}
