% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectNonWear.R
\name{detectNonWear}
\alias{detectNonWear}
\title{Detects Non Wear Periods}
\usage{
detectNonWear(data, sf, epoch, sdThreshold = 13/1000)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{sf}{Numeric indicating the sampling frequency in Hertz.}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sdThreshold}{Threshold to define nonwear time in g units (default 13 mgs).}
}
\value{
Numeric vector of length = nrow(data) indicating wear (0) and nonwear (1).
}
\description{
Function to detect non-wear periods based on the SD of the vector magnitude
of the raw accelerations. In brief, it looks for 30-min windows in which the
SD of the vector magnitude is lower than a given threshold (default = 0.013 G
units). Additionally, if short non-wear periods (max 30 min) are surrounded with
at least 30% of their duration by non-wear periods, then these short wear periods
are converted to non-wear (additional non-wear).
}
