% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSleep.R
\name{detectSleep}
\alias{detectSleep}
\title{Detects Sleep Periods}
\usage{
detectSleep(data, ts, epoch, sf, start_time)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{ts}{Data frame with ts object from \link{extractFeatures}}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{sf}{Numeric with the sampling frequency in Hz.}
}
\description{
Detects Sleep Periods
}
