% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresTrost2018.R
\name{featuresTrost2018}
\alias{featuresTrost2018}
\title{Calculates Features Proposed By Trost 2018}
\usage{
featuresTrost2018(
  data,
  vm,
  epoch,
  overlap,
  sf,
  lowerBound = 0.25,
  upperBound = 5
)
}
\arguments{
\item{data}{Matrix with 3 columns containing the raw acceleration for X, Y,
and Z axes in G units.}

\item{vm}{Numeric with vector magnitude of the raw accelerations.}

\item{epoch}{Numeric with the epoch length in seconds.}

\item{overlap}{Whether to overlap windows in the features calculation.}

\item{sf}{Numeric with the sampling frequency in Hz.}

\item{lowerBound}{Lower bound for the frequency filter in Hz.}

\item{upperBound}{Upper bound for the frequency filter in Hz.}
}
\value{
Matrix with features per epoch as proposed in Trost 2018.
}
\description{
Calculates Features Proposed By Trost 2018
}
\references{
Trost et al., Medicine and Science in Sport and Exercise 2018;50(3): 634-641
}
