% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate summary spreadsheet from processed
%   actigraph data.R, R/getbout.R
\name{getbout}
\alias{getbout}
\title{Title}
\usage{
getbout(
  x,
  boutduration,
  boutcriter = 0.8,
  closedbout = FALSE,
  bout.metric = 1,
  ws3 = 15
)

getbout(
  x,
  boutduration,
  boutcriter = 0.8,
  closedbout = FALSE,
  bout.metric = 1,
  ws3 = 15
)
}
\arguments{
\item{ws3}{}
}
\description{
Title

Title
}
