% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInfoClassifier.R
\name{GetInfoClassifier}
\alias{GetInfoClassifier}
\title{Get Information To Apply Classifiers}
\usage{
GetInfoClassifier(classifier)
}
\arguments{
\item{classifier}{Character specifying the classifier to be used
(available options are:
Preschool Wrist Random Forest Free Living,
Preschool Hip Random Forest Free Living,
Preschool Hip Random Forest Free Living Lag-Lead,
Preschool Wrist Random Forest Free Living Lag-Lead,
School age Wrist Random Forest, School age Hip Random Forest,
Adult Wrist RF Trost,
Adult women Wrist RF Ellis,
Adult women Hip RF Ellis,
Thigh Decision Tree)}
}
\value{
List containing epoch, rfmodel, and hmmmodel
}
\description{
Gets the required information to run a classifier (i.e., epoch and models)
}
