% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbed.R
\name{inbed}
\alias{inbed}
\title{Calculates inbed times.}
\usage{
inbed(
  angle,
  k = 60,
  perc = 0.1,
  inbedthreshold = 15,
  bedblocksize = 30,
  outofbedsize = 60,
  ws3 = 5
)
}
\arguments{
\item{angle}{Numeric with angle calculated.}

\item{k}{Interval for rolling window.}

\item{perc}{Percentage to establish threshold for inbed time definition.}

\item{inbedthreshold}{Threshold for inbed time definition.}

\item{bedblocksize}{Minimum time interval to define a bed block.}

\item{outofbedsize}{Maximum gap allowed in bed blocks.}

\item{ws3}{Epoch size.}
}
\value{
List with identificators of inbed time.
}
\description{
Calculates inbed times.
}
