% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbed.R
\name{inbed}
\alias{inbed}
\title{Calculates inbed times.}
\usage{
inbed(
  angle,
  k = 60,
  perc = 0.1,
  inbedthreshold = 15,
  bedblocksize = 30,
  outofbedsize = 60,
  ws3 = 5
)
}
\arguments{
\item{angle}{Numeric with angle calculated.}

\item{k}{Interval for rolling window.}

\item{perc}{Percentage to establish threshold for inbed time definition.}

\item{inbedthreshold}{Threshold for inbed time definition.}

\item{bedblocksize}{Minimum time interval to define a bed block.}

\item{outofbedsize}{Maximum gap allowed in bed blocks.}

\item{ws3}{Epoch size.}
}
\value{
List with identificators of inbed time.
}
\description{
Calculates inbed times.
}
\references{
GGIR R Package. doi: 10.1123/jmpb.2018-0063

van Hees VT, Sabia S, Jones SE, Wood AR, Anderson KN, Kivimäki M,
Frayling TM, Pack AI, Bucan M, Trenell MI, Mazzotti DR, Gehrman PR, Singh-Manoux BA,
Weedon MN. Estimating sleep parameters using an accelerometer without sleep diary.
Sci Rep. 2018 Aug 28;8(1):12975. doi: 10.1038/s41598-018-31266-z.

Ahmadi MN, Nathan N, Sutherland R, Wolfenden L, Trost SG. Non-wear
or sleep? Evaluation of five non-wear detection algorithms for raw accelerometer
data. J Sports Sci. 2020 Feb;38(4):399-404. doi: 10.1080/02640414.2019.1703301
}
\author{
Matthew N. Ahmadi <matthew.ahmadi@sydney.edu.au>
}
