% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAccFile.R
\name{ReadAndCalibrate}
\alias{ReadAndCalibrate}
\title{Read Accelerometer Raw Data File and Extract Calibration Coefficients}
\usage{
ReadAndCalibrate(
  file,
  sf,
  blocksize,
  blocknumber,
  inspectfileobject,
  PreviousEndPage,
  PreviousLastValue,
  PreviousLastTime,
  isLastBlock,
  do.calibration,
  iteration,
  epoch,
  S,
  verbose
)
}
\arguments{
\item{epoch}{}
}
\value{
List containing header info, accelerometer raw data, start time and ID
for the file.
}
\description{
Function aimed to read accelerometer raw data. At the moment,
the function supports ActiGraph gt3x and csv data, GENEActiv bin
and csv data
}
