% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndCalibrate.R
\name{ReadAndCalibrate}
\alias{ReadAndCalibrate}
\title{Read Accelerometer Raw Data File and Extract Calibration Coefficients}
\usage{
ReadAndCalibrate(
  file,
  sf,
  blocksize,
  blocknumber,
  inspectfileobject,
  PreviousEndPage,
  PreviousLastValue,
  PreviousLastTime,
  isLastBlock,
  do.calibration,
  iteration,
  epoch,
  S,
  verbose
)
}
\arguments{
\item{file}{Path to the file to be read.}

\item{sf}{Number with the sampling frequency in the recording.}

\item{blocksize}{Size of chunk to be read in pages as extracted from \code{GGIR::get_nw_clip_block_params}.}

\item{blocknumber}{Number of chunk of data to be read.}

\item{inspectfileobject}{Inspect file object extracted from \code{GGIR::g.inspectfile}.}

\item{PreviousEndPage}{Last page number read in the previous iteration.}

\item{PreviousLastValue}{Last acceleration value(s) read in the previous iteration.}

\item{PreviousLastTime}{Last timestamp read in the previous iteration.}

\item{isLastBlock}{Logical indicating if this is the last chunk of data to be read in the file.}

\item{do.calibration}{Logical indicating whether to calibrate the data or not.}

\item{iteration}{Number indicating the iteration in the \code{runActimetric} function.}

\item{epoch}{Number with the desired epoch length for the aggregation in seconds.}

\item{S}{Leftover data from the previous iteration, to be appended to the current chunk of data being read.}

\item{verbose}{Logical indicating whether to print progress messages in the console.}
}
\value{
List containing header info, accelerometer raw data, start time and ID
for the file.
}
\description{
Function aimed to read accelerometer raw data. At the moment,
the function supports ActiGraph gt3x and csv data, GENEActiv bin
and csv data
}
