% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runActimetric.R
\name{runActimetric}
\alias{runActimetric}
\title{Classifies Accelerometer Data into Physical Activity Types}
\usage{
runActimetric(
  input_directory = NULL,
  output_directory = NULL,
  studyname = "actimetric",
  do.calibration = TRUE,
  do.sleep = TRUE,
  do.nonwear = TRUE,
  do.enmo = TRUE,
  do.actilifecounts = FALSE,
  do.actilifecountsLFE = FALSE,
  classifier = NULL,
  boutdur = c(10),
  boutcriter = 0.8,
  boutmaxgap = 1,
  n_valid_hours = 0,
  n_valid_hours_awake = 0,
  n_valid_hours_nighttime = 0,
  visualreport = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input_directory}{Character with full path to directory containing raw
accelerometer data. At the moment the package accepts ActiGraph gt3x,
axivity cwa, and GENEActiv bin files. The \code{input_directory} should only
contain accelerometer files, they can be organized in subdirectories within
the input directory if needed.}

\item{output_directory}{Character with full path to directory where the reports
and visualizations are to be stored.}

\item{studyname}{Character (default = "actimetric") defining the name of the study.
To be used to give a name to the output directory generated.}

\item{do.calibration}{Logical (default = TRUE) indicating whether the accelerometer
raw signal should be calibrated relative to local gravitational acceleration.}

\item{do.sleep}{Logical (default = TRUE) indicating whether sleep parameters
should be calculated.}

\item{do.nonwear}{Logical (default = TRUE) indicating whether non-wear time
should be derived.}

\item{do.enmo}{Logical (default = TRUE) indicating whether ENMO should be
calculated.}

\item{do.actilifecounts}{Logical (default = FALSE) indicating whether activity
counts should be calculated.}

\item{do.actilifecountsLFE}{Logical (default = FALSE) indicating whether activity
counts using the low-frequency extension filter should be calculated.}

\item{classifier}{Character (default = NULL) indicating the classifier to be used
(available options are:
Preschool Wrist Random Forest Free Living,
Preschool Hip Random Forest Free Living,
Preschool Hip Random Forest Free Living Lag-Lead,
Preschool Wrist Random Forest Free Living Lag-Lead,
School age Wrist Random Forest, School age Hip Random Forest,
Adult Wrist RF Trost,
Adult women Wrist RF Ellis,
Adult women Hip RF Ellis)}

\item{boutdur}{Numeric vector (default = c(10)) indicating the bout durations over which calculate bouts of behaviors}

\item{boutcriter}{Numeric (default = 0.8) indicating the proportion of the bout duration that should be classified in a given behavior to consider a bout}

\item{boutmaxgap}{Integer (default = 1) with maximum consecutive gap length allowed in bout calculation.}

\item{n_valid_hours}{Numeric (default = 0) with minimum number of absolute valid hours in the day to consider it a valid day for the person-level aggregates.}

\item{n_valid_hours_awake}{Numeric (default = 0) with minimum number of absolute valid awake hours in the day to consider it a valid day for the person-level aggregates.}

\item{n_valid_hours_nighttime}{Numeric (default = 0) with minimum number of absolute valid nighttime hours in the day to consider it a valid day for the person-level aggregates.}

\item{visualreport}{Logical (default = FALSE) indicating whether visualizations should be run and stored in the results folder.}

\item{overwrite}{Logical (default = FALSE) indicating whether the existing milestone
data should be overwritten.}

\item{verbose}{Logical (default = TRUE) indicating whether progress messages should
be printed in the console.}
}
\value{
Function does not return anything, it only generates the reports and
visualizations in the \code{output_directory}.
}
\description{
This is the central function of the actimetric package.
}
\references{
GGIR R Package. doi: 10.1123/jmpb.2018-0063
}
